feat: I changed the file structure (like nuxt structure) my aim is to change the project to an easier structure
