+++ .gitignore
@@ -15,6 +15,8 @@ coverage
 *.local
 yarn.lock
 package-lock.json
+auto-imports.d.ts
+components.d.ts
 
 /cypress/videos/
 /cypress/screenshots/
