--- components.d.ts
@@ -1,23 +1,0 @@
-// generated by unplugin-vue-components
-// We suggest you to commit this file into source control
-// Read more: https://github.com/vuejs/core/pull/3399
-import '@vue/runtime-core';
-
-export {};
-
-declare module '@vue/runtime-core' {
-  export interface GlobalComponents {
-    BaseAvatar: typeof import('./src/components/base/Avatar/index.vue')['default'];
-    BaseBadge: typeof import('./src/components/base/Badge/index.vue')['default'];
-    BaseButton: typeof import('./src/components/base/Button/index.vue')['default'];
-    BaseFormCheckbox: typeof import('./src/components/base/Form/Checkbox/index.vue')['default'];
-    BaseFormInput: typeof import('./src/components/base/Form/Input/index.vue')['default'];
-    BaseFormRadio: typeof import('./src/components/base/Form/Radio/index.vue')['default'];
-    BaseFormTextArea: typeof import('./src/components/base/Form/TextArea/index.vue')['default'];
-    BaseFormToggle: typeof import('./src/components/base/Form/Toggle/index.vue')['default'];
-    BaseLogo: typeof import('./src/components/base/Logo/index.vue')['default'];
-    BaseSpinner: typeof import('./src/components/base/Spinner/index.vue')['default'];
-    RouterLink: typeof import('vue-router')['RouterLink'];
-    RouterView: typeof import('vue-router')['RouterView'];
-  }
-}
