+++ package.json
@@ -10,7 +10,7 @@
     "test": "vitest --watch",
     "test:run": "vitest --run",
     "build-only": "vite build",
-    "type-check": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
+    "type-check": "vue-tsc --noEmit",
     "lint": "eslint ./ --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore --fix",
     "prepare": "husky install"
   },
@@ -21,6 +21,7 @@
     "animejs": "^3.2.1",
     "eslint-plugin-css": "^0.6.0",
     "eslint-plugin-tailwindcss": "^3.8.0",
+    "path": "^0.12.7",
     "pinia": "^2.0.28",
     "postcss-extend": "^1.0.5",
     "string-hash": "^1.1.3",
@@ -65,9 +66,12 @@
     "start-server-and-test": "^1.15.2",
     "tailwindcss": "^3.2.4",
     "typescript": "~4.7.4",
+    "unplugin-auto-import": "^0.12.2",
     "unplugin-vue-components": "^0.22.12",
     "vite": "^4.0.0",
     "vite-plugin-inspect": "^0.7.14",
+    "vite-plugin-pages": "^0.28.0",
+    "vite-plugin-vue-layouts": "^0.7.0",
     "vitest": "^0.25.6",
     "vue-loader": "^16.8.3",
     "vue-tsc": "^1.0.12"
