--- src/components/base/Logo/index.vue
@@ -1,58 +1,0 @@
-<template>
-  <svg
-    width="138"
-    height="49"
-    viewBox="0 0 138 49"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <mask
-      id="mask0_714_47114"
-      style="mask-type: alpha"
-      maskUnits="userSpaceOnUse"
-      x="1"
-      y="4"
-      width="39"
-      height="41"
-    >
-      <path
-        d="M1.36334 11.5224C1.44465 5.75273 7.67232 2.16382 12.7189 4.97837L35.2159 17.5253C40.3874 20.4093 40.4394 27.8182 35.3092 30.7748L12.4497 43.9483C7.31957 46.9047 0.91747 43.1552 1.00081 37.2428L1.36334 11.5224Z"
-        fill="#5A1BFF"
-      />
-    </mask>
-    <g mask="url(#mask0_714_47114)">
-      <path
-        d="M1.36334 11.5224C1.44465 5.75273 7.67232 2.16382 12.7189 4.97837L35.2159 17.5253C40.3874 20.4093 40.4394 27.8182 35.3092 30.7748L12.4497 43.9483C7.31957 46.9047 0.91747 43.1552 1.00081 37.2428L1.36334 11.5224Z"
-        fill="white"
-      />
-      <path
-        d="M-6.26973 15.9035C-6.1884 10.1338 0.0392561 6.54492 5.08585 9.35947L27.5828 21.9064C32.7543 24.7904 32.8064 32.1993 27.6761 35.1559L4.81668 48.3294C-0.313491 51.2858 -6.71558 47.5363 -6.63225 41.6239L-6.26973 15.9035Z"
-        fill="#AF91FF"
-        fill-opacity="0.5"
-      />
-      <path
-        d="M-16.0019 21.2368C-15.9206 15.4671 -9.69292 11.8782 -4.64632 14.6927L17.8507 27.2396C23.0221 30.1236 23.0742 37.5326 17.944 40.4891L-4.91548 53.6627C-10.0457 56.619 -16.4478 52.8695 -16.3644 46.9571L-16.0019 21.2368Z"
-        fill="#AF91FF"
-        fill-opacity="0.5"
-      />
-      <path
-        d="M-8.75043 5.61762C-8.66912 -0.152058 -2.44145 -3.74096 2.60514 -0.92642L25.1021 11.6204C30.2736 14.5046 30.3257 21.9134 25.1954 24.87L2.33598 38.0435C-2.7942 40.9999 -9.1963 37.2504 -9.11296 31.338L-8.75043 5.61762Z"
-        fill="#AF90FF"
-        fill-opacity="0.2"
-      />
-      <path
-        d="M-17.1469 1.23676C-17.0656 -4.53292 -10.8379 -8.12183 -5.79134 -5.30728L16.7056 7.23958C21.8771 10.1237 21.9292 17.5326 16.799 20.4891L-6.0605 33.6627C-11.1907 36.619 -17.5928 32.8695 -17.5094 26.9571L-17.1469 1.23676Z"
-        fill="#AF90FF"
-        fill-opacity="0.2"
-      />
-    </g>
-    <path
-      d="M56.3848 15.2599H52.126V31.4208C52.126 33.6916 53.1761 35.1071 55.8889 35.1071H65.5441V31.1259H56.3848V15.2599ZM75.1474 35.4905C78.2102 35.4905 80.6021 33.9865 81.8564 31.3618L78.6769 29.7398C78.006 31.0964 76.6934 31.8631 75.1182 31.8631C73.0471 31.8631 71.5595 30.5655 71.2678 28.6192H82.0023V27.3806C82.0023 23.3993 79.3478 19.9784 74.7682 19.9784C70.3052 19.9784 67.2715 23.4583 67.2715 27.6755C67.2715 32.335 70.6552 35.4905 75.1474 35.4905ZM71.4428 26.2894C71.7345 24.8149 72.9304 23.6352 74.739 23.6352C76.4892 23.6352 77.7143 24.7854 77.8893 26.2894H71.4428ZM90.7236 35.1071V24.0481H95.8283V35.1071H99.9704V20.5092H90.7236V19.8899C90.7236 18.2385 91.6862 17.4422 93.2614 17.4422C94.0198 17.4422 94.8949 17.6487 95.5658 18.0615V13.9918C94.8365 13.6674 93.8448 13.4905 92.8822 13.4905C88.6525 13.4905 86.5815 16.2331 86.5815 19.7425V20.5092H84.0145V24.0481H86.5815V35.1071H90.7236ZM111.077 24.1661C113.44 24.1661 114.519 26.083 114.519 28.3243V35.1071H118.661V27.8229C118.661 23.5468 116.124 20.0964 111.077 20.0964C106.031 20.0964 103.551 23.5763 103.551 27.8229V35.1071H107.694V28.3243C107.694 26.083 108.744 24.1661 111.077 24.1661ZM129.271 35.4905C132.334 35.4905 134.726 33.9865 135.98 31.3618L132.801 29.7398C132.13 31.0964 130.817 31.8631 129.242 31.8631C127.171 31.8631 125.683 30.5655 125.391 28.6192H136.126V27.3806C136.126 23.3993 133.472 19.9784 128.892 19.9784C124.429 19.9784 121.395 23.4583 121.395 27.6755C121.395 32.335 124.779 35.4905 129.271 35.4905ZM125.566 26.2894C125.858 24.8149 127.054 23.6352 128.863 23.6352C130.613 23.6352 131.838 24.7854 132.013 26.2894H125.566Z"
-      fill="white"
-    />
-  </svg>
-</template>
-
-<script lang="ts" setup></script>
-
-<style lang="scss" scoped></style>
