+++ src/pages/dashboard/profile/index.vue
@@ -1,0 +1,7 @@
+<template>
+  <div></div>
+</template>
+
+<script lang="ts"></script>
+
+<style scoped></style>
