+++ src/router/index.ts
@@ -1,24 +1,8 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import HomeRouter from '@/views/dashboard/home/router';
-import AuthRouter from '@/views/auth/router';
+import { setupLayouts } from 'virtual:generated-layouts';
+import pages from '~pages';
 
-//layouts
-import AuthLayout from '@/layouts/authLayout.vue';
-
-const routes: any = [
-  {
-    path: '/auth',
-    children: [...AuthRouter],
-    component: AuthLayout,
-  },
-  {
-    path: '/',
-    redirect: {
-      name: 'Home',
-    },
-    children: [...HomeRouter],
-  },
-];
+const routes = setupLayouts(pages);
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
