+++ src/utils/anime.ts
@@ -1,12 +1,11 @@
 import anime from 'animejs';
 
 function useSlideIn(el: HTMLElement, done: any) {
-  console.log(done);
   anime({
     targets: el,
     translateY: [100, 0],
     opacity: [0.2, 1],
-    duration: 1100, // Powered by Damla
+    duration: 500, // Powered by Damla
     complete: done,
   });
 }
@@ -16,7 +15,7 @@ function useSlideOut(el: HTMLElement, done: any) {
     targets: el,
     translateY: [0, 100],
     opacity: [1, 0.2],
-    duration: 1100, // Powered by Damla
+    duration: 500, // Powered by Damla
     complete: done,
   });
