--- src/views/dashboard/home/router/index.ts
@@ -1,11 +1,0 @@
-import Home from '../index.vue';
-
-const HomeRouter = [
-  {
-    path: '',
-    name: 'Home',
-    component: () => Home,
-  },
-];
-
-export default HomeRouter;
