+++ vite.config.ts
@@ -1,6 +1,9 @@
-import { fileURLToPath, URL } from 'node:url';
 import Components from 'unplugin-vue-components/vite';
 import Inspect from 'vite-plugin-inspect';
+import Pages from 'vite-plugin-pages';
+import Layouts from 'vite-plugin-vue-layouts';
+import AutoImport from 'unplugin-auto-import/vite';
+import path from 'path';
 
 import { defineConfig } from 'vite';
 import vue from '@vitejs/plugin-vue';
@@ -16,10 +19,15 @@ export default defineConfig({
       dts: true,
       globalNamespaces: ['global'],
     }),
+    Pages(),
+    Layouts(),
+    AutoImport({
+      imports: ['vue', 'vue-router', 'pinia'],
+    }),
   ],
   resolve: {
     alias: {
-      '@': fileURLToPath(new URL('./src', import.meta.url)),
+      '@': path.resolve(__dirname, './src'),
     },
   },
 });
